pyproject.toml
[project]
name = "report-tool"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.26.0",
    "python-docx>=1.2.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.2",
    "httpx>=0.27.0",
    "watchdog>=6.0.0",
]

[tool.uv]
python-preference = "only-managed"
